safari-600.7-branch/LayoutTests/http/tests/security/contentSecurityPolicy/frame-src-blocked-expected.txt

CONSOLE MESSAGE: Refused to frame 'http://127.0.0.1:8000/security/contentSecurityPolicy/resources/alert-fail.html' because it violates the following Content Security Policy directive: "frame-src 'none'".
